Moved notifications lambda pages to settings
