+++ portal-ui/src/screens/Console/Console.tsx
@@ -39,7 +39,7 @@ import api from "../../common/api";
 import Account from "./Account/Account";
 import Users from "./Users/Users";
 import Groups from "./Groups/Groups";
-import ListNotificationEndpoints from "./NotificationEndopoints/ListNotificationEndpoints";
+import ListNotificationEndpoints from "./Configurations/NotificationEndpoints/ListNotificationEndpoints";
 import ConfigurationMain from "./Configurations/ConfigurationMain";
 import WebhookPanel from "./Configurations/ConfigurationPanels/WebhookPanel";
 import ListTenants from "./Tenants/ListTenants/ListTenants";
@@ -247,10 +247,6 @@ const Console = ({
     {
       component: Logs,
       path: "/logs",
-    },
-    {
-      component: ListNotificationEndpoints,
-      path: "/notification-endpoints",
     },
     {
       component: ConfigurationMain,
