+++ portal-ui/src/screens/Console/Menu/Menu.tsx
@@ -278,14 +278,6 @@ const Menu = ({ userLoggedIn, classes, pages }: IMenuProps) => {
       group: "Admin",
       type: "item",
       component: NavLink,
-      to: "/notification-endpoints",
-      name: "Lambda Notifications",
-      icon: <LambdaNotificationsIcon />,
-    },
-    {
-      group: "Admin",
-      type: "item",
-      component: NavLink,
       to: "/settings",
       name: "Settings",
       icon: <ConfigurationsListIcon />,
